move: ui컴포넌트 서버 및 클라이언트로 분리 (#173)
